remove caching client *and* decorators
