+++ frontend/openchat-agent/src/services/userIndex/userIndex.client.ts
@@ -1,5 +1,4 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
-/* eslint-disable @typescript-eslint/ban-types */
+import { groupBy } from "../../utils/list";
 import type { Identity } from "@dfinity/agent";
 import { Principal } from "@dfinity/principal";
 import { idlFactory, UserIndexService } from "./candid/idl";
@@ -19,6 +18,7 @@ import type {
     SetUserUpgradeConcurrencyResponse,
     ReferralLeaderboardRange,
     ReferralLeaderboardResponse,
+    PartialUserSummary,
 } from "openchat-shared";
 import { CandidService } from "../candidService";
 import {
@@ -35,18 +35,19 @@ import {
     referralLeaderboardResponse,
     userRegistrationCanisterResponse,
 } from "./mappers";
-import {
-    getUsersDecorator,
-    payForDiamondMembershipDecorator,
-    setUsernameDecorator,
-} from "./decorators";
 import { apiOptional } from "../common/chatMappers";
 import type { AgentConfig } from "../../config";
+import {
+    getCachedUsers,
+    setCachedUsers,
+    setUserDiamondStatusToTrueInCache,
+    setUsernameInCache,
+} from "../../utils/userCache";
 
 export class UserIndexClient extends CandidService {
     private userIndexService: UserIndexService;
 
-    private constructor(identity: Identity, config: AgentConfig) {
+    constructor(identity: Identity, config: AgentConfig) {
         super(identity);
 
         this.userIndexService = this.createServiceClient<UserIndexService>(
@@ -54,10 +55,6 @@ export class UserIndexClient extends CandidService {
             config.userIndexCanister,
             config
         );
-    }
-
-    static create(identity: Identity, config: AgentConfig): UserIndexClient {
-        return new UserIndexClient(identity, config);
     }
 
     getCurrentUser(): Promise<CurrentUserResponse> {
@@ -86,8 +83,35 @@ export class UserIndexClient extends CandidService {
         );
     }
 
-    @getUsersDecorator()
-    getUsers(users: UsersArgs, _allowStale: boolean): Promise<UsersResponse> {
+    async getUsers(users: UsersArgs, allowStale: boolean): Promise<UsersResponse> {
+        const allUsers = users.userGroups.flatMap((g) => g.users);
+
+        const fromCache = await getCachedUsers(allUsers);
+
+        // We throw away all of the updatedSince values passed in and instead use the values from the cache, this
+        // ensures the cache is always correct and doesn't miss any updates
+        const args = this.buildGetUsersArgs(allUsers, fromCache, allowStale);
+
+        const response = await this.getUsersFromBackend(users);
+
+        const requestedFromServer = new Set<string>([...args.userGroups.flatMap((g) => g.users)]);
+
+        // We return the fully hydrated users so that it is not possible for the Svelte store to miss any updates
+        const mergedResponse = this.mergeGetUsersResponse(
+            allUsers,
+            requestedFromServer,
+            response,
+            fromCache
+        );
+
+        setCachedUsers(mergedResponse.users.filter(this.isUserSummary)).catch((err) =>
+            console.error("Failed to save users to the cache", err)
+        );
+
+        return mergedResponse;
+    }
+
+    private getUsersFromBackend(users: UsersArgs): Promise<UsersResponse> {
         const userGroups = users.userGroups.filter((g) => g.users.length > 0);
 
         if (userGroups.length === 0) {
@@ -109,6 +133,93 @@ export class UserIndexClient extends CandidService {
         );
     }
 
+    private buildGetUsersArgs(
+        users: string[],
+        fromCache: UserSummary[],
+        allowStale: boolean
+    ): UsersArgs {
+        const fromCacheGrouped = groupBy(fromCache, (u) => u.updated);
+        const fromCacheSet = new Set<string>(fromCache.map((u) => u.userId));
+
+        const args: UsersArgs = {
+            userGroups: [],
+        };
+
+        // Add the users not found in the cache and ask for all updates
+        const notFoundInCache = users.filter((u) => !fromCacheSet.has(u));
+        if (notFoundInCache.length > 0) {
+            args.userGroups.push({
+                users: notFoundInCache,
+                updatedSince: BigInt(0),
+            });
+        }
+
+        if (!allowStale) {
+            // Add the users found in the cache but only ask for updates since the date they were last updated in the cache
+            for (const [updatedSince, users] of fromCacheGrouped) {
+                args.userGroups.push({
+                    users: users.map((u) => u.userId),
+                    updatedSince,
+                });
+            }
+        }
+
+        return args;
+    }
+
+    // Merges the cached values into the response
+    private mergeGetUsersResponse(
+        allUsers: string[],
+        requestedFromServer: Set<string>,
+        response: UsersResponse,
+        fromCache: UserSummary[]
+    ): UsersResponse {
+        if (fromCache.length === 0) {
+            return response;
+        }
+
+        const fromCacheMap = new Map<string, UserSummary>(fromCache.map((u) => [u.userId, u]));
+        const responseMap = new Map<string, PartialUserSummary>(
+            response.users.map((u) => [u.userId, u])
+        );
+
+        const users: PartialUserSummary[] = [];
+
+        for (const userId of allUsers) {
+            const cached = fromCacheMap.get(userId);
+            const userResponse = responseMap.get(userId);
+
+            if (userResponse !== undefined) {
+                users.push({
+                    ...userResponse,
+                    username: userResponse.username ?? cached?.username,
+                    blobReference: userResponse.blobReference ?? cached?.blobReference,
+                });
+            } else if (cached !== undefined) {
+                if (requestedFromServer.has(userId)) {
+                    // If this user was requested from the server but wasn't included in the response, then that means
+                    // our cached copy is up to date.
+                    users.push({
+                        ...cached,
+                        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                        updated: response.serverTimestamp!,
+                    });
+                } else {
+                    users.push(cached);
+                }
+            }
+        }
+
+        return {
+            serverTimestamp: response.serverTimestamp,
+            users,
+        };
+    }
+
+    private isUserSummary(user: PartialUserSummary): user is UserSummary {
+        return user.username !== undefined;
+    }
+
     checkUsername(username: string): Promise<CheckUsernameResponse> {
         const args = {
             username: username,
@@ -120,14 +231,18 @@ export class UserIndexClient extends CandidService {
         );
     }
 
-    @setUsernameDecorator()
-    setUsername(_userId: string, username: string): Promise<SetUsernameResponse> {
+    setUsername(userId: string, username: string): Promise<SetUsernameResponse> {
         return this.handleResponse(
             this.userIndexService.set_username({
                 username: username,
             }),
             setUsernameResponse
-        );
+        ).then((res) => {
+            if (res === "success") {
+                setUsernameInCache(userId, username);
+            }
+            return res;
+        });
     }
 
     suspendUser(userId: string, reason: string): Promise<SuspendUserResponse> {
@@ -154,9 +269,8 @@ export class UserIndexClient extends CandidService {
         return this.handleResponse(this.userIndexService.mark_suspected_bot({}), () => "success");
     }
 
-    @payForDiamondMembershipDecorator()
     payForDiamondMembership(
-        _userId: string,
+        userId: string,
         token: Cryptocurrency,
         duration: DiamondMembershipDuration,
         recurring: boolean,
@@ -170,7 +284,12 @@ export class UserIndexClient extends CandidService {
                 expected_price_e8s: expectedPriceE8s,
             }),
             payForDiamondMembershipResponse
-        );
+        ).then((res) => {
+            if (res.kind === "success") {
+                setUserDiamondStatusToTrueInCache(userId);
+            }
+            return res;
+        });
     }
 
     setUserUpgradeConcurrency(value: number): Promise<SetUserUpgradeConcurrencyResponse> {
